+++ app/popups/notification.js
@@ -41,6 +41,11 @@ ipcRenderer.on("notification-details", (_, message, fs, customCSS, lvl) => {
     }
     fontsize = fs
     document.body.style.fontSize = `${fontsize}px`
+    if (!document.getElementById("custom-styling")) {
+        const styleElement = document.createElement("style")
+        styleElement.id = "custom-styling"
+        document.head.append(styleElement)
+    }
     const customStyle = document.getElementById("custom-styling")
     if (customStyle) {
         customStyle.textContent = customCSS