+++ app/renderer/command.js
@@ -433,8 +433,7 @@ const colorscheme = (name = null, trailingArgs = null) => {
         css = readFile(joinPath(appData(), `colors/${name}.css`))
     }
     if (!css) {
-        css = readFile(joinPath(__dirname,
-            "../colors", `${name}.css`))
+        css = readFile(joinPath(__dirname, `../colors/${name}.css`))
     }
     if (!css) {
         notify(`Cannot find colorscheme '${name}'`, "warn")
@@ -443,9 +442,14 @@ const colorscheme = (name = null, trailingArgs = null) => {
     if (name === "default") {
         css = ""
     }
-    const customStyleEl = document.getElementById("custom-styling")
-    if (customStyleEl) {
-        customStyleEl.textContent = css
+    if (!document.getElementById("custom-styling")) {
+        const styleElement = document.createElement("style")
+        styleElement.id = "custom-styling"
+        document.head.append(styleElement)
+    }
+    const customStyle = document.getElementById("custom-styling")
+    if (customStyle) {
+        customStyle.textContent = css
     }
     ipcRenderer.send("set-custom-styling", getSetting("guifontsize"), css)
     const {setCustomStyling} = require("./settings")